app-1/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
chromadb==0.5.3
sentence-transformers==2.7.0
python-dotenv==1.0.0
pypdf==3.17.1
werkzeug==3.0.1
openai==1.3.0

# IBM Watson / WatsonX integration (optional - only needed if using IBM provider)
ibm-watsonx-ai>=0.1.0
langchain-ibm>=0.1.0
